% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{handle_model}
\alias{handle_model}
\alias{handle_model.default}
\alias{handle_model.workflow}
\title{Wrapper function for creating model handler function}
\usage{
handle_model(x, req, ...)

\method{handle_model}{default}(x, req, ...)

\method{handle_model}{workflow}(x, req, ...)
}
\arguments{
\item{x}{A trained model created with \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{req}{A POST request, with a \code{body}.}

\item{...}{Other arguments passed from \code{\link[=pr_model]{pr_model()}}.}
}
\description{
Wrapper function for creating model handler function
}
