% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{handle_model.workflow}
\alias{handle_model.workflow}
\title{Wrapper function for creating model handler function}
\usage{
\method{handle_model}{workflow}(x, ...)
}
\arguments{
\item{x}{A trained model created with \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{...}{Other arguments passed from \code{\link[=pr_model]{pr_model()}}.}
}
\description{
Wrapper function for creating model handler function
}
