% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops_handlers.R
\name{handler_startup.workflow}
\alias{handler_startup.workflow}
\alias{handler_predict.workflow}
\title{Model handler functions for API endpoint}
\usage{
\method{handler_startup}{workflow}(modelops, ...)

\method{handler_predict}{workflow}(modelops, ...)
}
\arguments{
\item{modelops}{A deployable model object created with \code{\link[modelops:modelops]{modelops()}}}

\item{...}{Other arguments passed to \code{predict()}, such as prediction \code{type}}
}
\description{
Model handler functions for API endpoint
}
