% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-model.R
\name{model_pinner.workflow}
\alias{model_pinner.workflow}
\title{Wrapper function for pinning a model to a board of models}
\usage{
\method{model_pinner}{workflow}(x, ...)
}
\arguments{
\item{x}{A trained model created with \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{...}{Other arguments passed from \code{\link[=pin_model]{pin_model()}}.}
}
\description{
Wrapper function for pinning a model to a board of models
}
