% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-model.R
\name{pin_model}
\alias{pin_model}
\title{Pin a trained modeling workflow to a board of models}
\usage{
pin_model(
  board,
  model,
  model_id,
  type = "rds",
  desc = NULL,
  metadata = NULL,
  versioned = TRUE
)
}
\arguments{
\item{board}{A pin board, created by \code{\link[pins:board_folder]{pins::board_folder()}},
\code{\link[pins:board_rsconnect]{pins::board_rsconnect()}}, or other \code{board_} function from the pins package.}

\item{model}{A trained modeling \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{model_id}{Model ID or name.}

\item{type}{Defaults to \code{"rds"}, which is appropriate for most models. (See
\code{\link[pins:pin_read]{pins::pin_write()}} for other options.)}

\item{desc}{A text description of the model; most important for shared
boards so that others can understand what the model is. If omitted,
tidymodelsdeploy will generate a brief description of the contents.}

\item{metadata}{A list containing additional metadata to store with the pin.
When retrieving the pin, this will be stored in the \code{user} key, to
avoid potential clashes with the metadata that pins itself uses.}

\item{versioned}{Should the model object be versioned? Defaults to \code{TRUE}.}
}
\description{
Use \code{pin_model()} to pin a trained model workflow to a board of models,
along with an input prototype for new data and model metadata.
}
\details{
This function creates \emph{two} pins on the specified \code{board}, one for
the model object itself and one for the model's input data prototype. The
naming convention for the model data prototype uses the \code{model_id} plus the
suffix "_ptype".
}
\examples{

library(pins)
model_board <- board_temp()

library(parsnip)
library(workflows)
rf_spec <- rand_forest(mode = "regression") \%>\%
    set_engine("ranger")

mtcars_wf <- workflow() \%>\%
    add_model(rf_spec) \%>\%
    add_formula(mpg ~ .) \%>\%
    fit(data = mtcars)


model_board \%>\% pin_model(mtcars_wf, "mtcars_ranger")
# notice two pins created
model_board

}
