% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pr_model}
\alias{pr_model}
\title{Add a POST endpoint to a Plumber router using a pinned model workflow object}
\usage{
pr_model(pr, board, model_id, type = NULL, path = "/predict", ...)
}
\arguments{
\item{pr}{A Plumber router, such as from \code{\link[plumber:pr]{plumber::pr()}}.}

\item{board}{A board containing models pinned via \code{\link{pin_model}}.}

\item{model_id}{Model ID or name.}

\item{type}{A single character value or \code{NULL}. Possible values
are "numeric", "class", "prob", "conf_int", "pred_int", "quantile", "time",
"hazard", "survival", or "raw". When \code{NULL}, \code{predict()} will choose an
appropriate value based on the model's mode.}

\item{path}{The endpoint path}

\item{...}{Other arguments passed to \code{\link[plumber:pr_handle]{plumber::pr_post()}}.}
}
\description{
Models that have been pinned to a board with \code{\link{pin_model}} can be added to a
Plumber router as a POST handler. The argument \code{type} specifies what kind of
predictions the handler will return.
}
\details{
The model's input data prototype is added to the handler to ensure
predictions can be made correctly, and packages needed for prediction are
loaded.
}
\examples{

library(pins)
model_board <- board_temp()

library(parsnip)
library(workflows)
rf_spec <- rand_forest(mode = "regression") \%>\%
    set_engine("ranger")

mtcars_wf <- workflow() \%>\%
    add_model(rf_spec) \%>\%
    add_formula(mpg ~ .) \%>\%
    fit(data = mtcars)

model_board \%>\%
    pin_model(mtcars_wf, "mtcars_ranger")

library(plumber)
pr() \%>\%
    pr_model(model_board, "mtcars_ranger")
## next, pipe to `pr_run()`

}
