% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-model.R
\name{vetiver_model.workflow}
\alias{vetiver_model.workflow}
\title{Create a vetiver object for deployment of a trained tidymodels workflow}
\usage{
\method{vetiver_model}{workflow}(
  model,
  model_name,
  board,
  ...,
  desc = NULL,
  metadata = list(),
  save_ptype = TRUE,
  versioned = NULL
)
}
\arguments{
\item{model}{A trained model created with \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{model_name}{Model name or ID.}

\item{board}{A pin board to store and version the \code{model}, created by
\code{\link[pins:board_folder]{pins::board_folder()}}, \code{\link[pins:board_rsconnect]{pins::board_rsconnect()}}, or other \code{board_}
function from the pins package.}

\item{...}{Other method-specific arguments passed to \code{\link[vetiver:vetiver_create_ptype]{vetiver_slice_zero()}}
to compute an input data prototype.}

\item{desc}{A text description of the model, most important for shared
boards so that others can understand what the model is. If omitted,
a brief description of the contents will be generated.}

\item{metadata}{A list containing additional metadata to store with the pin.
When retrieving the pin, this will be stored in the \code{user} key, to
avoid potential clashes with the metadata that pins itself uses.}

\item{save_ptype}{Should an input data prototype be stored with the model?
The options are \code{TRUE} (the default, which stores a zero-row slice of the
training data), \code{FALSE} (no input data prototype for checking), or a
dataframe.}

\item{versioned}{Should the model object be versioned? The default, \code{NULL},
will use the default for \code{board}.}
}
\description{
A \code{\link[vetiver:vetiver_model]{vetiver::vetiver_model()}} object collects the information needed to
store, version, and deploy a trained model.
}
