% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-handlers.R
\name{handler_startup.workflow}
\alias{handler_startup.workflow}
\alias{handler_predict.workflow}
\title{Model handler functions for API endpoint}
\usage{
\method{handler_startup}{workflow}(vetiver_model, ...)

\method{handler_predict}{workflow}(vetiver_model, ...)
}
\arguments{
\item{vetiver_model}{A deployable \code{\link[vetiver:vetiver_model]{vetiver_model()}} object}

\item{...}{Other arguments passed to \code{predict()}, such as prediction \code{type}}
}
\description{
Model handler functions for API endpoint
}
