% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{model_pinner}
\alias{pin_model}
\alias{pr_model}
\alias{handle_model}
\alias{api_spec}
\alias{model_endpoint}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{modelops}{\code{\link[modelops]{api_spec}}, \code{\link[modelops]{handle_model}}, \code{\link[modelops]{model_endpoint}}, \code{\link[modelops]{model_pinner}}, \code{\link[modelops]{pin_model}}, \code{\link[modelops]{pr_model}}}
}}

