% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{modelops}
\alias{modelops_pin_write}
\alias{modelops_pr_predict}
\alias{handler_startup}
\alias{handler_predict}
\alias{api_spec}
\alias{modelops_endpoint}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{modelops}{\code{\link[modelops]{api_spec}}, \code{\link[modelops]{handler_predict}}, \code{\link[modelops:handler_predict]{handler_startup}}, \code{\link[modelops]{modelops}}, \code{\link[modelops]{modelops_endpoint}}, \code{\link[modelops]{modelops_pin_write}}, \code{\link[modelops]{modelops_pr_predict}}}
}}

